requirements.txt
pandas
numpy
matplotlib
tensorflow
scikit-learn
openpyxl
python-dotenv
pillow
psutil
gputil
